requirements_frozen.txt
-e git+ssh://git@github.com/swansonk14/admet_ai.git@4d025463b6de10c5cb2f481eb5d0694a53be0bbb#egg=admet_ai
aimsim_core==2.2.3
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
astartes==1.3.3
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
chemprop==2.2.2
click==8.3.1
ConfigArgParse==1.7.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
descriptastorus==2.8.0
dill==0.4.1
docstring_parser==0.17.0
et_xmlfile==2.0.0
executing==2.2.1
filelock==3.24.1
Flask==3.1.2
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
gunicorn==25.1.0
idna==3.11
iniconfig==2.3.0
ipdb==0.13.13
ipython==9.10.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lightning==2.6.1
lightning-utilities==0.15.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mhfp==1.9.6
mordredcommunity==2.0.7
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.2
openpyxl==3.1.5
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079068/work
padelpy==0.1.16
pandas==3.0.0
pandas-flavor==0.8.1
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.1
PyYAML==6.0.3
rdkit==2025.9.5
rich==14.3.2
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
setuptools==80.10.2
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
torch==2.10.0
torchmetrics==1.8.2
tqdm==4.67.3
traitlets==5.14.3
typed-argument-parser==1.11.0
typing-inspect==0.9.0
typing_extensions==4.15.0
wcwidth==0.6.0
Werkzeug==3.1.5
wheel==0.46.3
xarray==2026.2.0
yarl==1.22.0
